+++ src/main/java/com/kkarimi/eventmanagement/datashipper/ChangeHistoryRepository.java
@@ -1,6 +1,0 @@
-package com.kkarimi.eventmanagement.datashipper;
-
-import org.springframework.data.mongodb.repository.MongoRepository;
-
-interface ChangeHistoryRepository extends MongoRepository<ChangeHistoryDocument, String> {
-}