+++ src/main/java/com/kkarimi/eventmanagement/datashipper/package-info.java
@@ -1,2 +1,0 @@
-@org.springframework.modulith.ApplicationModule(displayName = "Data Shipper")
-package com.kkarimi.eventmanagement.datashipper;