+++ src/main/java/com/kkarimi/eventmanagement/eventhistory/EventHistoryRepository.java
@@ -1,0 +1,6 @@
+package com.kkarimi.eventmanagement.eventhistory;
+
+import org.springframework.data.mongodb.repository.MongoRepository;
+
+interface EventHistoryRepository extends MongoRepository<EventHistoryDocument, String> {
+}