+++ src/main/java/com/kkarimi/eventmanagement/eventhistory/package-info.java
@@ -1,0 +1,2 @@
+@org.springframework.modulith.ApplicationModule(displayName = "Event History")
+package com.kkarimi.eventmanagement.eventhistory;