+++ src/main/java/com/kkarimi/eventmanagement/registration/package-info.java
@@ -1,5 +1,5 @@
 @org.springframework.modulith.ApplicationModule(
         displayName = "Registration",
-        allowedDependencies = {"events", "attendees", "notifications", "datashipper"}
+        allowedDependencies = {"events", "attendees", "notifications", "eventhistory"}
 )
 package com.kkarimi.eventmanagement.registration;